+++ tests/cli/workdir-legacy/src/index.js
@@ -1,0 +1,3 @@
+export { default as Required } from "./Required.svelte";
+export { default as SlotsNamed } from "./SlotsNamed.svelte";
+export { default as TypedProps } from "./TypedProps.svelte";