+++ tests/cli/workdir-svelte5/src/index.js
@@ -1,0 +1,5 @@
+export { default as EdgeCases } from "./EdgeCases.svelte";
+export { default as NoInterface } from "./NoInterface.svelte";
+export { default as RegionBadge } from "./RegionBadge.svelte";
+export { default as SimpleDefaults } from "./SimpleDefaults.svelte";
+export { default as Tooltip } from "./Tooltip.svelte";